NIFI-2501: - Ensuring users can access the controller service list regardless of permissions on the corresponding process group or controller.
