+++ nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-client-dto/src/main/java/org/apache/nifi/web/api/entity/ControllerConfigurationEntity.java
@@ -20,11 +20,8 @@ import com.wordnik.swagger.annotations.ApiModelProperty;
 import org.apache.nifi.web.api.dto.ControllerConfigurationDTO;
 import org.apache.nifi.web.api.dto.PermissionsDTO;
 import org.apache.nifi.web.api.dto.RevisionDTO;
-import org.apache.nifi.web.api.dto.util.TimeAdapter;
 
 import javax.xml.bind.annotation.XmlRootElement;
-import javax.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
-import java.util.Date;
 
 /**
  * A serialized representation of this class can be placed in the entity body of a request or response to or from the API. This particular entity holds a reference to a ControllerConfigurationDTO.
@@ -32,7 +29,6 @@ import java.util.Date;
 @XmlRootElement(name = "controllerConfigurationEntity")
 public class ControllerConfigurationEntity extends Entity implements Permissible<ControllerConfigurationDTO> {
 
-    private Date currentTime;
     private ControllerConfigurationDTO controllerConfiguration;
     private RevisionDTO revision;
     private PermissionsDTO permissions;
@@ -87,18 +83,4 @@ public class ControllerConfigurationEntity extends Entity implements Permissible
         this.permissions = permissions;
     }
 
-    /**
-     * @return current time on the server
-     */
-    @XmlJavaTypeAdapter(TimeAdapter.class)
-    @ApiModelProperty(
-            value = "The current time on the system."
-    )
-    public Date getCurrentTime() {
-        return currentTime;
-    }
-
-    public void setCurrentTime(Date currentTime) {
-        this.currentTime = currentTime;
-    }
 }
