--- nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-web/nifi-web-ui/src/main/webapp/WEB-INF/partials/canvas/process-group-details.jsp
@@ -1,39 +1,0 @@
-<%--
- Licensed to the Apache Software Foundation (ASF) under one or more
-  contributor license agreements.  See the NOTICE file distributed with
-  this work for additional information regarding copyright ownership.
-  The ASF licenses this file to You under the Apache License, Version 2.0
-  (the "License"); you may not use this file except in compliance with
-  the License.  You may obtain a copy of the License at
-
-      http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
---%>
-<%@ page contentType="text/html" pageEncoding="UTF-8" session="false" %>
-<div id="process-group-details" class="hidden medium-dialog">
-    <div class="dialog-content">
-        <div class="setting">
-            <div class="setting-name">Name</div>
-            <div class="setting-field">
-                <span id="read-only-process-group-name"></span>
-            </div>
-        </div>
-        <div class="setting">
-            <div class="setting-name">Id</div>
-            <div class="setting-field">
-                <span id="read-only-process-group-id"></span>
-            </div>
-        </div>
-        <div class="setting">
-            <div class="setting-name">Comments</div>
-            <div class="setting-field">
-                <div id="read-only-process-group-comments"></div>
-            </div>
-        </div>
-    </div>
-</div>
