+++ nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-web/nifi-web-ui/src/main/webapp/js/nf/canvas/controllers/nf-ng-canvas-global-menu-controller.js
@@ -123,9 +123,7 @@ nf.ng.Canvas.GlobalMenuCtrl = function (serviceProvider) {
                  * Launch the settings shell.
                  */
                 launch: function () {
-                    if (nf.Common.canAccessController()) {
-                        nf.Settings.showSettings();
-                    }
+                    nf.Settings.showSettings();
                 }
             }
         };
